// src/types/FilterCheckbox.ts
export type FilterCheckbox = {
  data: {
    description?: string
    text: string
    /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
    value: any
  }[]
  id: string
  title: string
  type: 'Checkbox'
  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
  value: any[]
}